email_service/requirements.txt
confluent-kafka==2.3.0
python-dotenv==1.0.1
prometheus_client==0.20.0
